Add TrajOpt planner config
This commit adds the files needed for running the trajopt planner example that is found in the
[moveit_tutorials](https://ros-planning.github.io/moveit_tutorials/doc/trajopt_planner/trajopt_planner_tutorial.html).
This planner is not included into `move_group.launch` by default but can be invoked via the `pipeline` argument.
This was done since it is not yet officially released (see https://github.com/ros-planning/moveit/issues/1467).