Rework collision models for hand
... utilizing a new feature of MoveIt / SRDF
Now, it's possible to disable collision for a specific collision name by default.
The definition of specific collision pairs will override this default.

This allows disabling collision for the coarse hand geometries
and enabling it for selected pairs of links only (for self-collision checking).

On the other hand, the detailed hand collision geometries are enabled by default,
but selectively disabled for self-collision checking for all arm links.